Muzzle logs should be logged using the io.opentelemetry.* logger name
